backend/requirements.txt
bs4
docopt
docopt
filelock
flask
flask-cors
loguru
pillow
pydantic
pydantic
pyexcel-xls
pyexcel>=0.6.7
pyyaml
requests
tqdm
xxhash
